fixed server crash when reading long LLDP IDs

+++ src/server/include/nms_topo.h
@@ -305,7 +305,7 @@ void AddLLDPNeighbors(Node *node, LinkLayerNeighbors *nbs);
 void AddNDPNeighbors(Node *node, LinkLayerNeighbors *nbs);
 void AddCDPNeighbors(Node *node, LinkLayerNeighbors *nbs);
 void AddSTPNeighbors(Node *node, LinkLayerNeighbors *nbs);
-void BuildLldpId(int type, const BYTE *data, size_t length, TCHAR *id, int idLen);
+void BuildLldpId(int type, const BYTE *data, size_t length, TCHAR *id, size_t idLen);
 
 void BridgeMapPorts(SNMP_Transport *transport, InterfaceList *ifList);
